refactor: extract action cards from AgentWorkspace
